update codes, upgrade BLINKER PRINT codes
